+++ src/robotsix_central_deploy/lifecycle/backends/__init__.py
@@ -10,12 +10,13 @@ from .base import ExecutionBackend
 from .docker_cli import DockerBackend
 from .docker_sdk import DockerSdkBackend
 from .noop import NoopBackend
-from ._util import collect_protected_image_refs
+from ._util import PruneImagesResult, collect_protected_image_refs
 
 __all__ = [
     "ExecutionBackend",
     "DockerBackend",
     "DockerSdkBackend",
     "NoopBackend",
+    "PruneImagesResult",
     "collect_protected_image_refs",
 ]